Set reboot annotation on the bare metal host
When we need to remediate the node that currently runs the remediation
controller, possible situation when we will poweroff the host, and
until all controller will restart on other node, the reboot will failed
on the timeout and the node will stay in poweroff state forever, because
first reboot failed and the others will succeed because host already has
`spec.Online` parameter equals to `false`, so the controller will think that
an user power offed it.
To prevent it we will set annotation on the `BareMetalHost` resource, before
poweroff operation and remove it after poweron, in this case, if the first
reboot will fail, the one that will follow it, will recognize annotation and will continue
the reboot process.

Signed-off-by: Artyom Lukianov <alukiano@redhat.com>