Pass controllers images as ENV variables under the operator
Possible situation when the name of the image can be different under the u/s and d/s,
in this case we want to have flexibility to specify image names for all controllers.

In case if you did not specify MHC, MDB or MR image names, the operator will try to
build it, from the operator image repository, component name and operator version.

- update csv, manifest and template generatros to get images variables
- update the operator, to get image names from the ENV variables

Signed-off-by: Artyom Lukianov <alukiano@redhat.com>